+++ app/Http/Requests/Product/CreateProductRequest.php
@@ -29,7 +29,7 @@ class CreateProductRequest extends FormRequest
     {
         return [
             'name' => ['string','required','unique:products,name'],
-            'price'=> ['required','integer','min:1'],
+            'price'=> ['required','numeric','min:1'],
             // 'status' => ['nullable', new Enum(StatusEnum::class)],
             'path' => ["required","image", "mimes:jpeg,jpg,png,gif,svg,webp", "max:5120"],
         ];
